# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BootstrapStatus {
  needsSetup: boolean;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  ADMIN: 'ADMIN',
  USER: 'USER',
} as const;

export interface User {
  id: number;
  nombre: string;
  email: string;
  role: UserRole;
  isActive: boolean;
  creadoEn: string;
}

export interface UserCredentialsInput {
  /** @minLength 1 */
  nombre: string;
  email: string;
  /** @minLength 12 */
  password: string;
  role?: UserRole;
}

export interface UserUpdateInput {
  /** @minLength 1 */
  nombre: string;
  email: string;
  role: UserRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface PasswordInput {
  /** @minLength 12 */
  password: string;
}

export interface UserStatusInput {
  isActive: boolean;
}

export interface SessionResponse {
  authenticated: boolean;
  user: User | null;
}

export interface AuthenticatedSessionResponse {
  authenticated: true;
  user: User;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  ACTIVA: 'ACTIVA',
  BLOQUEADA: 'BLOQUEADA',
  FINALIZADA: 'FINALIZADA',
} as const;

export type CoilStatus = typeof CoilStatus[keyof typeof CoilStatus];


export const CoilStatus = {
  DISPONIBLE: 'DISPONIBLE',
  EN_FÁBRICA: 'EN FÁBRICA',
} as const;

export type Camisa = typeof Camisa[keyof typeof Camisa];


export const Camisa = {
  NUMBER_400: 400,
  NUMBER_475: 475,
  NUMBER_520: 520,
  '22-6-22': '22-6-22',
  '21-8-21': '21-8-21',
  '40-6-40': '40-6-40',
  '40-8-40': '40-8-40',
  '47-5-47': '47-5-47',
  '47-8-47': '47-8-47',
  '52-8-52': '52-8-52',
} as const;

export type Material = typeof Material[keyof typeof Material];


export const Material = {
  OPP: 'OPP',
  OPP_RECICLADO: 'OPP RECICLADO',
} as const;

export interface CoilCharacteristics {
  ancho: number;
  micras: number;
  camisa: Camisa;
  material: Material;
}

export type OrderInput = CoilCharacteristics & {
  /** @exclusiveMinimum 0 */
  metrosNecesarios: number;
};

export interface OrderBlockInput {
  blocked: boolean;
}

export interface OrderFinalizeInput {
  nota?: string;
  /** Meter deficit the caller is showing to the operator. When it no longer matches the freshly computed deficit (e.g. Nexus grouped a pedido while the dialog was open) the request is rejected with 409 and the current deficit, so the operator confirms real numbers. */
  faltantesEsperados?: number;
  /** Finalize regardless of the deficit check. Used to confirm after a 409, when the operator has already seen the up-to-date deficit. */
  forzar?: boolean;
}

export interface OrderReopenInput {
  motivo?: string;
}

export type OrderReorderStatus = typeof OrderReorderStatus[keyof typeof OrderReorderStatus];


export const OrderReorderStatus = {
  ACTIVA: 'ACTIVA',
  BLOQUEADA: 'BLOQUEADA',
} as const;

export interface OrderReorderInput {
  estado?: OrderReorderStatus;
  /** @minItems 1 */
  orderIds: number[];
}

export type OrderOrigin = typeof OrderOrigin[keyof typeof OrderOrigin];


export const OrderOrigin = {
  MANUAL: 'MANUAL',
  GESTION_PEDIDOS: 'GESTION_PEDIDOS',
} as const;

export interface RelatedPedido {
  id: number;
  pedidoId: string;
  numeroPedidoCliente: string;
  metros: number;
  vinculadoEn: string;
}

export interface NexusOrderInput {
  eventId: string;
  /** @minLength 1 */
  pedidoId: string;
  /** @minLength 1 */
  numeroPedidoCliente: string;
  /** @exclusiveMinimum 0 */
  metros: number;
  /** @exclusiveMinimum 0 */
  bobinaMadre: number;
  /** @minLength 1 */
  camisa: string;
  /** @minLength 1 */
  tipoMaterial: string;
  /** @exclusiveMinimum 0 */
  micras: number;
}

export type NexusOrderResponseAction = typeof NexusOrderResponseAction[keyof typeof NexusOrderResponseAction];


export const NexusOrderResponseAction = {
  ORDER_CREATED: 'ORDER_CREATED',
  ORDER_UPDATED: 'ORDER_UPDATED',
  ALREADY_PROCESSED: 'ALREADY_PROCESSED',
} as const;

export interface NexusOrderResponse {
  success: boolean;
  action: NexusOrderResponseAction;
  orderId: number;
  totalMetros: number;
  eventId: string;
}

export interface ProductionOrder {
  id: number;
  ancho: number;
  micras: number;
  camisa: string;
  material: string;
  metrosNecesarios: number;
  metrosFabricados: number;
  metrosPendientes: number;
  estado: OrderStatus;
  origen: OrderOrigin;
  pedidosRelacionados: RelatedPedido[];
  creadoEn: string;
  /** @nullable */
  finalizadaEn: string | null;
  /** @nullable */
  nota: string | null;
}

export type ProductionOrderEventAccion = typeof ProductionOrderEventAccion[keyof typeof ProductionOrderEventAccion];


export const ProductionOrderEventAccion = {
  BLOQUEADA: 'BLOQUEADA',
  DESBLOQUEADA: 'DESBLOQUEADA',
  FINALIZADA_MANUAL: 'FINALIZADA_MANUAL',
  FINALIZADA_AUTO: 'FINALIZADA_AUTO',
  REABIERTA: 'REABIERTA',
  PEDIDO_AGRUPADO: 'PEDIDO_AGRUPADO',
} as const;

export interface ProductionOrderEvent {
  id: number;
  ordenId: number;
  /** @nullable */
  usuarioId: number | null;
  /**
     * Null when the action was taken by an automated process.
     * @nullable
     */
  usuarioNombre: string | null;
  accion: ProductionOrderEventAccion;
  /** @nullable */
  detalle: string | null;
  creadoEn: string;
}

export interface ManufacturedCoilInput {
  ordenId: number;
  /**
     * @minimum 100
     * @maximum 25000
     */
  metros: number;
}

export interface RemnantInput {
  ancho: number;
  micras: number;
  camisa: string;
  material: string;
  metros: number;
}

export interface CoilMaterialInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  material: string;
}

export interface UpdateCoilInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  material?: string;
  /** @minimum 1 */
  metros?: number;
  /**
     * @minLength 1
     * @maxLength 60
     */
  camisa?: string;
}

export interface CoilStockAssignment {
  ordenId: number;
  metros: number;
  origen: string;
  asignadoEn: string;
}

export type CoilTipo = typeof CoilTipo[keyof typeof CoilTipo];


export const CoilTipo = {
  BOBINA: 'BOBINA',
  RESTO: 'RESTO',
} as const;

export type Coil = CoilCharacteristics & ({
  id: number;
  tipo: CoilTipo;
  metros: number;
  estado: CoilStatus;
  /** @nullable */
  ordenId?: number | null;
  asignacion?: CoilStockAssignment | null;
  pedidosRelacionados?: RelatedPedido[];
  creadoEn: string;
  /** @nullable */
  movidoAFabricaEn?: string | null;
});

export interface InventorySummary {
  totalMetros: number;
  items: Coil[];
}

export type ListOrdersParams = {
status?: OrderStatus;
};

export type ListInventoryParams = {
status?: CoilStatus;
};