# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /integrations/gestion-pedidos/nexus-orders:
    post:
      operationId: processNexusOrder
      tags: [integrations]
      summary: Receive order from GestionPedidosPigmea
      description: Process and group customer order into a production order
      security:
        - IntegrationBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NexusOrderInput"
      responses:
        "200":
          description: Order updated or already processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NexusOrderResponse"
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NexusOrderResponse"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "503":
          description: Service integration not configured
components:
  securitySchemes:
    IntegrationBearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    NexusOrderInput:
      type: object
      required:
        [
          eventId,
          pedidoId,
          numeroPedidoCliente,
          metros,
          bobinaMadre,
          camisa,
          tipoMaterial,
          micras,
        ]
      properties:
        eventId:
          type: string
        pedidoId:
          type: string
          minLength: 1
        numeroPedidoCliente:
          type: string
          minLength: 1
        metros:
          type: number
          exclusiveMinimum: 0
        bobinaMadre:
          type: number
          exclusiveMinimum: 0
        camisa:
          type: string
          minLength: 1
        tipoMaterial:
          type: string
          minLength: 1
        micras:
          type: number
          exclusiveMinimum: 0
    NexusOrderResponse:
      type: object
      required: [success, action, orderId, totalMetros, eventId]
      properties:
        success:
          type: boolean
        action:
          type: string
          enum: [ORDER_CREATED, ORDER_UPDATED, ALREADY_PROCESSED]
        orderId:
          type: number
        totalMetros:
          type: number
        eventId:
          type: string
...